requirements.txt
openai-whisper
openai
python-dotenv
fastapi
uvicorn[standard]
python-multipart
aiofiles

# Database
sqlalchemy
psycopg2-binary
alembic
